feat: Enhance FAL AI provider with async job submission and status checking
