+++ llm_generate_job/models/__init__.py
@@ -1,3 +1,3 @@
 from . import llm_generate_job
-from . import fal_ai_provider
 from . import llm_thread
+from . import llm_provider
